+++ Source/swiftlint/Commands/AutoCorrectCommand.swift
@@ -38,6 +38,7 @@ struct AutoCorrectCommand: CommandProtocol {
         return configuration.visitLintableFiles(path: options.path, action: "Correcting",
                                                 quiet: options.quiet,
                                                 useScriptInputFiles: options.useScriptInputFiles,
+                                                forceExclude: options.forceExclude,
                                                 cache: cache, parallel: true) { linter in
             let corrections = linter.correct()
             if !corrections.isEmpty && !options.quiet {
@@ -53,7 +54,10 @@ struct AutoCorrectCommand: CommandProtocol {
             }
         }.flatMap { files in
             if !options.quiet {
-                queuedPrintError("Done correcting \(files.count) files!")
+                let pluralSuffix = { (collection: [Any]) -> String in
+                    return collection.count != 1 ? "s" : ""
+                }
+                queuedPrintError("Done correcting \(files.count) file\(pluralSuffix(files))!")
             }
             return .success(())
         }
@@ -65,16 +69,17 @@ struct AutoCorrectOptions: OptionsProtocol {
     let configurationFile: String
     let useScriptInputFiles: Bool
     let quiet: Bool
+    let forceExclude: Bool
     let format: Bool
     let cachePath: String
     let ignoreCache: Bool
     let useTabs: Bool
 
     // swiftlint:disable line_length
-    static func create(_ path: String) -> (_ configurationFile: String) -> (_ useScriptInputFiles: Bool) -> (_ quiet: Bool) -> (_ format: Bool) -> (_ cachePath: String) -> (_ ignoreCache: Bool) -> (_ useTabs: Bool) -> AutoCorrectOptions {
-        return { configurationFile in { useScriptInputFiles in { quiet in { format in { cachePath in { ignoreCache in { useTabs in
-            self.init(path: path, configurationFile: configurationFile, useScriptInputFiles: useScriptInputFiles, quiet: quiet, format: format, cachePath: cachePath, ignoreCache: ignoreCache, useTabs: useTabs)
-        }}}}}}}
+    static func create(_ path: String) -> (_ configurationFile: String) -> (_ useScriptInputFiles: Bool) -> (_ quiet: Bool) -> (_ forceExclude: Bool) -> (_ format: Bool) -> (_ cachePath: String) -> (_ ignoreCache: Bool) -> (_ useTabs: Bool) -> AutoCorrectOptions {
+        return { configurationFile in { useScriptInputFiles in { quiet in { forceExclude in { format in { cachePath in { ignoreCache in { useTabs in
+            self.init(path: path, configurationFile: configurationFile, useScriptInputFiles: useScriptInputFiles, quiet: quiet, forceExclude: forceExclude, format: format, cachePath: cachePath, ignoreCache: ignoreCache, useTabs: useTabs)
+        }}}}}}}}
     }
 
     static func evaluate(_ mode: CommandMode) -> Result<AutoCorrectOptions, CommandantError<CommandantError<()>>> {
@@ -84,6 +89,9 @@ struct AutoCorrectOptions: OptionsProtocol {
             <*> mode <| configOption
             <*> mode <| useScriptInputFilesOption
             <*> mode <| quietOption(action: "correcting")
+            <*> mode <| Option(key: "force-exclude",
+                               defaultValue: false,
+                               usage: "exclude files in config `excluded` even if their paths are explicitly specified")
             <*> mode <| Option(key: "format",
                                defaultValue: false,
                                usage: "should reformat the Swift files")